Finish to address Sylvain's comments
